Move FileLocalizer and Localizers to AppTemplate folder and namespace.
