+++ Sandra.UI/Localizers.cs
@@ -19,20 +19,12 @@
 **********************************************************************************/
 #endregion
 
-using Eutherion;
 using Eutherion.Localization;
 using Eutherion.Text.Json;
 using Eutherion.UIActions;
-using Eutherion.Utils;
 using Eutherion.Win.AppTemplate;
 using Eutherion.Win.Storage;
-using Eutherion.Win.UIActions;
-using System;
 using System.Collections.Generic;
-using System.Drawing;
-using System.Globalization;
-using System.IO;
-using System.Linq;
 
 namespace Sandra.UI
 {
@@ -86,228 +78,6 @@ namespace Sandra.UI
         internal static readonly LocalizedStringKey View = new LocalizedStringKey(nameof(View));
         internal static readonly LocalizedStringKey ZoomIn = new LocalizedStringKey(nameof(ZoomIn));
         internal static readonly LocalizedStringKey ZoomOut = new LocalizedStringKey(nameof(ZoomOut));
-    }
-
-    internal static class Localizers
-    {
-        private static readonly string NativeNameDescription
-            = "Name of the language in the language itself. This property is mandatory.";
-
-        public static readonly SettingProperty<string> NativeName = new SettingProperty<string>(
-            new SettingKey(SettingKey.ToSnakeCase(nameof(NativeName))),
-            TrimmedStringType.Instance,
-            new SettingComment(NativeNameDescription));
-
-        private static readonly string FlagIconFileDescription
-            = "File name of the flag icon to display in the menu.";
-
-        public static readonly SettingProperty<string> FlagIconFile = new SettingProperty<string>(
-            new SettingKey(SettingKey.ToSnakeCase(nameof(FlagIconFile))),
-            FileNameType.Instance,
-            new SettingComment(FlagIconFileDescription));
-
-        private static readonly string TranslationsDescription
-            = "List of translations.";
-
-        public static readonly SettingProperty<Dictionary<LocalizedStringKey, string>> Translations = new SettingProperty<Dictionary<LocalizedStringKey, string>>(
-            new SettingKey(SettingKey.ToSnakeCase(nameof(Translations))),
-            TranslationDictionaryType.Instance,
-            new SettingComment(TranslationsDescription));
-
-        public static SettingSchema CreateLanguageFileSchema()
-        {
-            return new SettingSchema(
-                NativeName,
-                FlagIconFile,
-                Translations);
-        }
-
-        public static IEnumerable<string> LocalizerKeyCandidates(CultureInfo uiCulture)
-        {
-            yield return uiCulture.NativeName;
-            yield return uiCulture.Name;
-            yield return uiCulture.ThreeLetterISOLanguageName;
-            yield return uiCulture.TwoLetterISOLanguageName;
-            yield return uiCulture.LCID.ToString();
-
-            // Recursively check parents, stop at the invariant culture which has no name.
-            if (uiCulture.Parent.Name.Length > 0)
-            {
-                foreach (var candidateKey in LocalizerKeyCandidates(uiCulture.Parent))
-                {
-                    yield return candidateKey;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Initializes the available localizers.
-        /// </summary>
-        /// <param name="defaultLangDirectory">
-        /// Path to the directory to scan for language files.
-        /// </param>
-        public static Dictionary<string, FileLocalizer> ScanLocalizers(string defaultLangDirectory)
-        {
-            var foundLocalizers = new Dictionary<string, FileLocalizer>(StringComparer.OrdinalIgnoreCase);
-
-            var languageFileSchema = CreateLanguageFileSchema();
-
-            try
-            {
-                DirectoryInfo defaultDir = new DirectoryInfo(defaultLangDirectory);
-
-                if (defaultDir.Exists)
-                {
-                    foreach (var fileInfo in defaultDir.EnumerateFiles("*.json"))
-                    {
-                        try
-                        {
-                            var languageFile = SettingsFile.Create(fileInfo.FullName, new SettingCopy(languageFileSchema));
-                            languageFile.Settings.TryGetValue(FlagIconFile, out string flagIconFileName);
-
-                            foundLocalizers.Add(
-                                Path.GetFileNameWithoutExtension(fileInfo.Name),
-                                new FileLocalizer(languageFile));
-                        }
-                        catch (Exception exc)
-                        {
-                            exc.Trace();
-                        }
-                    }
-                }
-            }
-            catch (Exception exc)
-            {
-                exc.Trace();
-            }
-
-            return foundLocalizers;
-        }
-
-        public static FileLocalizer BestFit(Dictionary<string, FileLocalizer> localizers)
-        {
-            if (localizers.Count > 0)
-            {
-                // Try keys based on CurrentUICulture.
-                foreach (var candidateKey in LocalizerKeyCandidates(CultureInfo.CurrentUICulture))
-                {
-                    if (localizers.TryGetValue(candidateKey, out FileLocalizer localizerMatch))
-                    {
-                        return localizerMatch;
-                    }
-                }
-
-                return localizers.First().Value;
-            }
-
-            return null;
-        }
-    }
-
-    /// <summary>
-    /// Apart from being a <see cref="Localizer"/>, contains properties
-    /// to allow construction of <see cref="UIAction"/> bindings and interact with settings.
-    /// </summary>
-    public sealed class FileLocalizer : Localizer, IWeakEventTarget
-    {
-        /// <summary>
-        /// Gets the settings file from which this localizer is loaded.
-        /// </summary>
-        public SettingsFile LanguageFile { get; }
-
-        /// <summary>
-        /// Gets the name of the language in the language itself, e.g. "English", "Español", "Deutsch", ...
-        /// </summary>
-        public string LanguageName { get; }
-
-        /// <summary>
-        /// Gets the file name without extension of the flag icon.
-        /// </summary>
-        public string FlagIconFileName { get; }
-
-        public Dictionary<LocalizedStringKey, string> Dictionary { get; private set; }
-
-        public FileLocalizer(SettingsFile languageFile)
-        {
-            LanguageFile = languageFile;
-            LanguageName = languageFile.Settings.GetValue(Localizers.NativeName);
-            FlagIconFileName = languageFile.Settings.TryGetValue(Localizers.FlagIconFile, out string flagIconFile) ? flagIconFile : string.Empty;
-            UpdateDictionary();
-        }
-
-        private void UpdateDictionary()
-        {
-            Dictionary = LanguageFile.Settings.TryGetValue(Localizers.Translations, out Dictionary<LocalizedStringKey, string> dict) ? dict : new Dictionary<LocalizedStringKey, string>();
-        }
-
-        public override string Localize(LocalizedStringKey localizedStringKey, string[] parameters)
-            => Dictionary.TryGetValue(localizedStringKey, out string displayText)
-            ? LocalizedString.ConditionalFormat(displayText, parameters)
-            : Default.Localize(localizedStringKey, parameters);
-
-        private DefaultUIActionBinding switchToLangUIActionBinding;
-
-        public DefaultUIActionBinding SwitchToLangUIActionBinding
-        {
-            get
-            {
-                if (switchToLangUIActionBinding == null)
-                {
-                    Image menuIcon = null;
-                    try
-                    {
-                        if (!string.IsNullOrEmpty(FlagIconFileName))
-                        {
-                            menuIcon = Image.FromFile(Path.Combine(Path.GetDirectoryName(LanguageFile.AbsoluteFilePath), FlagIconFileName));
-                        }
-                    }
-                    catch (Exception exc)
-                    {
-                        exc.Trace();
-                    }
-
-                    switchToLangUIActionBinding = new DefaultUIActionBinding(
-                        new UIAction(nameof(Localizers) + "." + LanguageName),
-                        new ImplementationSet<IUIActionInterface>
-                        {
-                            new ContextMenuUIActionInterface
-                            {
-                                MenuCaptionKey = LocalizedStringKey.Unlocalizable(LanguageName),
-                                MenuIcon = menuIcon,
-                            },
-                        });
-                }
-
-                return switchToLangUIActionBinding;
-            }
-        }
-
-        public UIActionState TrySwitchToLang(bool perform)
-        {
-            if (perform)
-            {
-                Current = this;
-                Session.Current.AutoSave.Persist(Session.Current.LangSetting, this);
-            }
-
-            return new UIActionState(UIActionVisibility.Enabled, Current == this);
-        }
-
-        public void EnableLiveUpdates()
-        {
-            // Can only happen after a message loop has been started.
-            LanguageFile.RegisterSettingsChangedHandler(Localizers.Translations, TranslationsChanged);
-        }
-
-        private void TranslationsChanged(object sender, EventArgs e)
-        {
-            UpdateDictionary();
-            NotifyChanged();
-        }
-
-        // TranslationsChanged only raises another WeakEvent so this does not indirectly leak.
-        // FileLocalizers do not go out of scope during the lifetime of the application.
-        bool IWeakEventTarget.IsDisposed => false;
     }
 
     internal sealed class BuiltInEnglishLocalizer : Localizer
@@ -463,109 +233,5 @@ namespace Sandra.UI
 
             return localizer.Localize(GetLocalizedStringKey(jsonErrorInfo.ErrorCode), jsonErrorInfo.Parameters);
         }
-    }
-
-    public sealed class TrimmedStringType : PType.Derived<string, string>
-    {
-        public static readonly PTypeErrorBuilder TrimmedStringTypeError
-            = new PTypeErrorBuilder(new LocalizedStringKey(nameof(TrimmedStringTypeError)));
-
-        public static TrimmedStringType Instance = new TrimmedStringType();
-
-        private TrimmedStringType() : base(PType.CLR.String) { }
-
-        public override string GetBaseValue(string value) => value;
-
-        public override Union<ITypeErrorBuilder, string> TryGetTargetValue(string value)
-            => !string.IsNullOrWhiteSpace(value)
-            ? ValidValue(value.Trim())
-            : InvalidValue(TrimmedStringTypeError);
-    }
-
-    public sealed class TranslationDictionaryType : PType.Derived<PMap, Dictionary<LocalizedStringKey, string>>
-    {
-        public static TranslationDictionaryType Instance = new TranslationDictionaryType();
-
-        private TranslationDictionaryType() : base(PType.Map) { }
-
-        public override PMap GetBaseValue(Dictionary<LocalizedStringKey, string> value)
-        {
-            Dictionary<string, PValue> dictionary = new Dictionary<string, PValue>();
-
-            if (value != null)
-            {
-                foreach (var kv in value)
-                {
-                    dictionary.Add(kv.Key.Key, new PString(kv.Value));
-                }
-            }
-
-            return new PMap(dictionary);
-        }
-
-        public override Union<ITypeErrorBuilder, Dictionary<LocalizedStringKey, string>> TryGetTargetValue(PMap value)
-        {
-            var dictionary = new Dictionary<LocalizedStringKey, string>();
-
-            foreach (var kv in value)
-            {
-                if (!PType.String.TryGetValidValue(kv.Value).IsOption2(out PString stringValue))
-                {
-                    return InvalidValue(new TranslationDictionaryTypeError(kv.Key, kv.Value));
-                }
-
-                dictionary.Add(new LocalizedStringKey(kv.Key), stringValue.Value);
-            }
-
-            return ValidValue(dictionary);
-        }
-    }
-
-    /// <summary>
-    /// Represents the result of a failed typecheck of <see cref="TranslationDictionaryType"/>.
-    /// </summary>
-    public class TranslationDictionaryTypeError : PValueVisitor<Localizer, string>, ITypeErrorBuilder
-    {
-        /// <summary>
-        /// Gets the translation key for this error message.
-        /// </summary>
-        public static readonly LocalizedStringKey LocalizedMessageKey = new LocalizedStringKey(nameof(TranslationDictionaryTypeError));
-
-        /// <summary>
-        /// Gets the key with the illegal value.
-        /// </summary>
-        public string Key { get; }
-
-        /// <summary>
-        /// Gets the illegal value.
-        /// </summary>
-        public PValue InvalidStringValue { get; }
-
-        /// <summary>
-        /// Initializes a new instance of <see cref="KeyedSetTypeError"/>.
-        /// </summary>
-        internal TranslationDictionaryTypeError(string key, PValue invalidStringValue)
-        {
-            Key = key;
-            InvalidStringValue = invalidStringValue;
-        }
-
-        /// <summary>
-        /// Gets the localized error message in the current language.
-        /// </summary>
-        public string GetLocalizedTypeErrorMessage(Localizer localizer, string propertyKey, string valueString)
-            => localizer.Localize(LocalizedMessageKey, new[]
-            {
-                propertyKey,
-                PTypeErrorBuilder.QuoteStringValue(Key),
-                Visit(InvalidStringValue, localizer)
-            });
-
-
-        public override string DefaultVisit(PValue value, Localizer localizer) => localizer.Localize(PType.JsonUndefinedValue);
-        public override string VisitBoolean(PBoolean value, Localizer localizer) => PTypeErrorBuilder.QuoteValue(JsonValue.BoolSymbol(value.Value));
-        public override string VisitInteger(PInteger value, Localizer localizer) => PTypeErrorBuilder.QuoteValue(value.Value.ToString(CultureInfo.InvariantCulture));
-        public override string VisitList(PList value, Localizer localizer) => localizer.Localize(PType.JsonArray);
-        public override string VisitMap(PMap value, Localizer localizer) => localizer.Localize(PType.JsonObject);
     }
 }
