+++ Sandra.UI/ToolForms.cs
@@ -23,6 +23,7 @@ using Eutherion;
 using Eutherion.Localization;
 using Eutherion.UIActions;
 using Eutherion.Utils;
+using Eutherion.Win.AppTemplate;
 using Eutherion.Win.Storage;
 using Eutherion.Win.UIActions;
 using System;
